refactor: 파일 경로 생성기를 유틸로 이동
