+++ backend/forgather/src/main/java/com/forgather/domain/space/util/FilePathGenerator.java
@@ -1,4 +1,4 @@
-package com.forgather.domain.space.service;
+package com.forgather.domain.space.util;
 
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
@@ -6,6 +6,9 @@ import org.springframework.web.multipart.MultipartFile;
 public class FilePathGenerator {
 
     private static final String CONTENTS_INNER_PATH = "contents";
+
+    private FilePathGenerator() {
+    }
 
     public static String generateContentsFilePath(String rootDirectory, String spaceCode, String fileName) {
         return String.format("%s/%s/%s/%s", rootDirectory, CONTENTS_INNER_PATH, spaceCode, fileName);
